Reorganized all stories.
