+++ src/stories/documentation/TwitchChat.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/addon-docs';
 
-<Meta title="TwitchChat/Documentation" />
+<Meta title="Documentation/TwitchChat" />
 
 # TwitchChat
